populate ClientType on RemoteControlClient

+++ RemoteControlClient.go
@@ -1,7 +1,7 @@
 package main
 
 type RemoteControlClient struct {
-	clientType            byte
+	clientType            int // either 256 for QU-Pad or 0 for QU-You
 	incomingSystemPackets chan SystemPacket
 	outgoingSystemPackets chan SystemPacket
 	incomingDSPPackets    chan DSPPacket
